@@ -1,50 +1,57 @@
 {
-  "name": "generator-lens-ext",
-  "version": "0.1.1",
-  "description": "Yeoman generator for Lens extension",
-  "copyright": "© 2020, Mirantis, Inc.",
+  "name": "@gharbi-yasmine/generator-freelens-ext",
+  "version": "0.0.1",
+  "description": "A generator for FreeLens extensions",
+  "copyright": "© 2025, Yasmine Gharbi © 2020, Mirantis, Inc. ",
   "license": "MIT",
+  "type": "module",
   "main": "generators/app/index.js",
   "author": {
-    "name": "Mirantis, Inc.",
-    "email": "info@k8slens.dev"
+    "name": "Yasmine Gharbi",
+    "email": "gyasmine29@gmail.com"
   },
   "scripts": {
     "test": "jest"
   },
+  "files": [
+    "generators"
+  ],
   "keywords": [
     "yeoman-generator",
     "lens",
     "kubernetes",
-    "extensions"
+    "extensions",
+    "yeoman-generator",
+    "freelensapp"
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/lensapp/generator-lens-ext.git"
+    "url": "https://github.com/GHARBIyasmine/generator-freelens-ext.git"
   },
   "bugs": {
-    "url": "https://github.com/lensapp/generator-lens-ext/issues"
+    "url": "https://github.com/GHARBIyasmine/generator-freelens-ext/issues"
   },
   "dependencies": {
-    "yeoman-generator": "^4.12.0",
-    "yosay": "^2.0.2"
+    "yeoman-generator": "^7.5.1",
+    "yosay": "^3.0.0"
   },
   "devDependencies": {
-    "@k8slens/extensions": "^6.1.9",
+    "@freelensapp/extensions": "^1.2.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^12.1.5",
-    "@types/jest": "^28.1.6",
-    "@types/node": "^16.18.2",
-    "@types/react": "^17.0.45",
-    "@typescript-eslint/eslint-plugin": "^5.45.0",
-    "@typescript-eslint/parser": "^5.45.0",
-    "eslint": "^8.28.0",
-    "jest": "^28.1.3",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.15.2",
+    "@types/react": "^17.0.85",
+    "@types/react-dom": "^17.0.26",
+    "mobx": "^6.10.2",
+		"mobx-react": "^7.6.0",
+    "css-loader": "^6.11.0",
+    "jest": "^29.7.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "typescript": "^4.9.3",
+    "typescript": "^5.8.2",
     "yeoman-assert": "^3.1.1",
-    "yeoman-test": "^3.0.0"
+    "yeoman-test": "^10.1.0"
   },
   "jest": {
     "modulePathIgnorePatterns": [
