@@ -1,14 +1,20 @@
-const Generator = require("yeoman-generator");
-const yosay = require("yosay");
-const path = require("path");
-const { version: generatorVersion } = require("../../package.json");
+import Generator from 'yeoman-generator';
+import yosay from 'yosay';
+import { fileURLToPath } from 'url';
+import path from 'path';
+import { readFile } from 'fs/promises';
+import { askForExtensionName, askForExtensionDescription, askForExtensionPublisher, askForGit, askForInstallDependencies, askForPackageManager } from './prompts.js';
+import { generateExtTs } from './generate-ext-ts.js';
+import { createSymlink } from './symlink.js';
+import dependencyVersions from './deps.js';
 
-const dependencyVersions = require("./deps");
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
-// TODO: add JS version
-// const commandjs = require("./generate-ext-js");
-const extTS = require("./generate-ext-ts");
-module.exports = class extends Generator {
+// Read package.json
+const packageJson = JSON.parse(await readFile(path.join(__dirname, '../../package.json'), 'utf-8'));
+const generatorVersion = packageJson.version;
+
+export default class extends Generator {
   constructor(args, opts) {
     super(args, opts);
     this.option("extensionType", { type: String });
@@ -22,8 +28,7 @@ module.exports = class extends Generator {
   }
 
   async initializing() {
-
-    this.log(yosay(`Welcome to the Lens Extension Generator v${generatorVersion}`));
+    this.log(yosay(`Welcome to the FreeLens Extension Generator v${generatorVersion}`));
 
     this.extensionConfig.dependencyVersions = dependencyVersions;
     this.extensionConfig.dep = function (name) {
@@ -36,91 +41,59 @@ module.exports = class extends Generator {
   }
 
   async prompting() {
-    const extensionGenerators = [
-      extTS
-    ];
-
-    // Ask for extension type
-    const choices = [];
-    for (const g of extensionGenerators) {
-      const name = g.name;
-      if (name) {
-        choices.push({ name, value: g.id });
-      }
-    }
-    this.extensionConfig.type = (await this.prompt({
-      type: "list",
-      name: "type",
-      message: "What type of extension do you want to create?",
-      pageSize: choices.length,
-      choices,
-    })).type;
-
-    this.extensionGenerator = extensionGenerators.find(g => g.id === this.extensionConfig.type);
-    try {
-      await this.extensionGenerator.prompting(this, this.extensionConfig);
-    } catch (e) {
-      this.abort = true;
-    }
-
+    await askForExtensionName(this, this.extensionConfig);
+    await askForExtensionDescription(this, this.extensionConfig);
+    await askForExtensionPublisher(this, this.extensionConfig);
+    await askForGit(this, this.extensionConfig);
+    await askForPackageManager(this, this.extensionConfig);
+    await askForInstallDependencies(this, this.extensionConfig);
   }
-  // Write files
+
   writing() {
     if (this.abort) {
       return;
     }
-    this.sourceRoot(path.join(__dirname, "./templates/" + this.extensionConfig.type));
-    return this.extensionGenerator.writing(this, this.extensionConfig);
+    generateExtTs(this, this.extensionConfig);
   }
 
-  // Installation
   install() {
     if (this.abort) {
       return;
     }
     if (this.extensionConfig.installDependencies) {
-      if (this.extensionConfig.pkgManager === "yarn") {
-        this.spawnCommand("yarn", [], {
-          cwd: this.destinationPath(`./${this.extensionConfig.name}`)
-        });
-      }
-      if (this.extensionConfig.pkgManager === "npm") {
-        this.spawnCommand("npm", ["install"], {
-          cwd: this.destinationPath(`./${this.extensionConfig.name}`)
-        });
-      }
+      this.spawnCommand(this.extensionConfig.pkgManager, ['install'], {
+        cwd: this.destinationPath(`./${this.extensionConfig.name}`)
+      });
     }
   }
 
-  // End
   end() {
     if (this.abort) {
       return;
     }
 
-    // Git init
     if (this.extensionConfig.gitInit) {
-      this.spawnCommandSync("git", ["init", "--quiet"], {
+      this.spawnCommandSync('git', ['init', '--quiet'], {
         cwd: this.destinationPath(`./${this.extensionConfig.name}`)
       });
-      this.spawnCommandSync("git", ["add", "-A"], {
+      this.spawnCommandSync('git', ['add', '-A'], {
         cwd: this.destinationPath(`./${this.extensionConfig.name}`)
       });
-      this.spawnCommandSync("git", ["commit", "-m", `Initial commit 🎉, generated by generator-lens-ext v${generatorVersion}`], {
+      this.spawnCommandSync('git', ['commit', '-m', `Initial commit 🎉, generated by generator-lens-ext v${generatorVersion}`], {
         cwd: this.destinationPath(`./${this.extensionConfig.name}`)
       });
     }
 
-    this.log("\r\n");
-    this.log("While developing your extension you should either work directly in '~/.k8slens/extensions' or symlink your working directory into that folder");
+    this.log('\r\n');
+    this.log("While developing your extension you should either work directly in '~/.freelens/extensions' or symlink your working directory into that folder");
 
-    this.log("\r\n");
+    this.log('\r\n');
     this.log(`Your extension "${this.extensionConfig.name}" has been created!`);
     this.log(`cd to ${this.extensionConfig.name}, and '${this.extensionConfig.pkgManager} start' to start the development.`);
-    this.log("\r\n");
+    this.log('\r\n');
 
-    this.log("Lens Extension Documentation https://docs.k8slens.dev/latest/extensions/.");
-    this.log("Join #lens-extensions on Lens Dev Slack http://k8slens.slack.com/");
-    this.log("\r\n");
+    this.log("Legacy Lens Extension Documentation https://docs.k8slens.dev/latest/extensions/.");
+    this.log("FreeLens GitHub https://github.com/freelensapp/freelens");
+    this.log('\r\n');
   }
-};
+}
